test(no-hungarian): Add comprehensive function call test cases
This commit adds extensive test cases to the `no-hungarian.test.ts` file to cover various scenarios of function calls with Hungarian notation parameters. The tests include:

- Function calls with method invocations
- String, array, and object method calls
- Callback and arrow function scenarios
- Method chaining
- External library function calls
- Function calls with different parameter types and notations

These test cases ensure that the `no-hungarian` rule correctly handles function calls and does not incorrectly flag valid use cases with Hungarian-like parameter names.
